Clean after rebase.
